refactoring

+++ src/main/java/com/traulko/project/service/impl/UserServiceImpl.java
@@ -9,7 +9,6 @@ import com.traulko.project.exception.ServiceException;
 import com.traulko.project.service.UserService;
 import com.traulko.project.util.CustomCipher;
 import com.traulko.project.util.mail.MailSender;
-import com.traulko.project.validator.ProjectValidator;
 import com.traulko.project.validator.UserValidator;
 
 import java.security.NoSuchAlgorithmException;
@@ -21,10 +20,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public boolean changePassword(String email, String password, String passwordRepeat) throws ServiceException {
-        UserValidator userValidator = new UserValidator();
         boolean result = false;
         try {
-            if (userValidator.isPasswordValid(password) && password.equals(passwordRepeat)) {
+            if (UserValidator.isEmailValid(email) && UserValidator.isPasswordValid(password)
+                    && password.equals(passwordRepeat)) {
                 CustomCipher cipher = new CustomCipher();
                 String encryptedPassword = cipher.encrypt(password);
                 result = userDao.changePassword(email, encryptedPassword);
@@ -33,6 +32,20 @@ public class UserServiceImpl implements UserService {
             throw new ServiceException("Error while changing password", e);
         }
         return result;
+    }
+
+    @Override
+    public Optional<User> findById(String id) throws ServiceException {
+        Optional<User> optionalUser = Optional.empty();
+        try {
+            if (UserValidator.isIdValid(id)) {
+                int userId = Integer.parseInt(id);
+                optionalUser = userDao.findById(userId);
+            }
+        } catch (DaoException e) {
+            throw new ServiceException("Error while finding user by id", e);
+        }
+        return optionalUser;
     }
 
     @Override
@@ -48,6 +61,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public Optional<User> findUserByAccessCode(String code, List<User> userList) throws ServiceException {
+        // TODO: 12.11.2020
         Optional<User> optionalUser = Optional.empty();
         CustomCipher cipher = new CustomCipher();
         try {
@@ -67,7 +81,9 @@ public class UserServiceImpl implements UserService {
     public boolean remove(String email) throws ServiceException {
         boolean isRemoved = false;
         try {
-            isRemoved = userDao.remove(email);
+            if (UserValidator.isEmailValid(email)) {
+                isRemoved = userDao.remove(email);
+            }
         } catch (DaoException e) {
             throw new ServiceException("Error while removing user by email", e);
         }
@@ -78,7 +94,9 @@ public class UserServiceImpl implements UserService {
     public boolean block(String email) throws ServiceException {
         boolean isBlocked = false;
         try {
-            isBlocked = userDao.block(email);
+            if (UserValidator.isEmailValid(email)) {
+                isBlocked = userDao.block(email);
+            }
         } catch (DaoException e) {
             throw new ServiceException("Error while blocking user", e);
         }
@@ -89,7 +107,9 @@ public class UserServiceImpl implements UserService {
     public boolean unblock(String email) throws ServiceException {
         boolean isUnblocked = false;
         try {
-            isUnblocked = userDao.unblock(email);
+            if (UserValidator.isEmailValid(email)) {
+                isUnblocked = userDao.unblock(email);
+            }
         } catch (DaoException e) {
             throw new ServiceException("Error while unblocking user", e);
         }
@@ -97,12 +117,15 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public boolean fillUpBalance(User user, String moneyAmount) throws ServiceException {
+    public boolean fillUpBalance(String userId, String moneyAmount) throws ServiceException {
         boolean result = false;
         try {
-            if (ProjectValidator.isCorrectDoubleValue(moneyAmount)) {
+            if (UserValidator.isIdValid(userId) && UserValidator.isPriceValid(moneyAmount)) {
                 double amount = Double.parseDouble(moneyAmount);
-                if (amount > 0) {
+                int userIdParsed = Integer.parseInt(userId);
+                Optional<User> optionalUser = userDao.findById(userIdParsed);
+                if (optionalUser.isPresent()) {
+                    User user = optionalUser.get();
                     double newBalance = user.getBalance() + amount;
                     user.setBalance(newBalance);
                     userDao.update(user);
@@ -131,12 +154,14 @@ public class UserServiceImpl implements UserService {
     public boolean activateUser(String email) throws ServiceException {
         boolean result = false;
         try {
-            Optional<User> optionalUser = userDao.findByEmail(email);
-            if (optionalUser.isPresent()) {
-                User user = optionalUser.get();
-                user.setStatus(User.Status.ENABLE);
-                if (userDao.update(user)) {
-                    result = true;
+            if (UserValidator.isEmailValid(email)) {
+                Optional<User> optionalUser = userDao.findByEmail(email);
+                if (optionalUser.isPresent()) {
+                    User user = optionalUser.get();
+                    user.setStatus(User.Status.ENABLE);
+                    if (userDao.update(user)) {
+                        result = true;
+                    }
                 }
             }
         } catch (DaoException e) {
@@ -150,7 +175,7 @@ public class UserServiceImpl implements UserService {
         try {
             return userDao.findAll();
         } catch (DaoException e) {
-            throw new ServiceException("Error while finding all users in batabase", e);
+            throw new ServiceException("Error while finding all users", e);
         }
     }
 
@@ -159,7 +184,7 @@ public class UserServiceImpl implements UserService {
         try {
             return userDao.findBySearchQuery(searchQuery);
         } catch (DaoException e) {
-            throw new ServiceException("Error while finding users by search query in batabase", e);
+            throw new ServiceException("Error while finding users by search query", e);
         }
     }
 
@@ -167,16 +192,15 @@ public class UserServiceImpl implements UserService {
     @Override
     public boolean isUserExists(String email, String password) throws ServiceException {
         boolean result = false;
-        if (UserValidator.isEmailValid(email) && UserValidator.isPasswordValid(password)) {
-            try {
+        try {
+            if (UserValidator.isEmailValid(email) && UserValidator.isPasswordValid(password)) {
                 CustomCipher cipher = new CustomCipher();
                 String encryptedPassword = cipher.encrypt(password);
                 Optional<User> optionalUser = userDao.findByEmailAndPassword(email, encryptedPassword);
-                result = optionalUser.isPresent() ? true : false;
-            } catch (DaoException | NoSuchAlgorithmException e) {
-                throw new ServiceException("Error while checking user " +
-                        "for presence in database", e);
+                result = optionalUser.isPresent();
             }
+        } catch (DaoException | NoSuchAlgorithmException e) {
+            throw new ServiceException("Error while checking user for presence in database", e);
         }
         return result;
     }
@@ -199,9 +223,7 @@ public class UserServiceImpl implements UserService {
                 user.setStatus(User.Status.NOT_CONFIRMED);
                 user.setRole(User.Role.USER);
                 user.setBalance(0);
-                if (userDao.add(user, encryptedPassword)) {
-                    result = true;
-                }
+                result = userDao.add(user, encryptedPassword);
             } catch (DaoException | NoSuchAlgorithmException e) {
                 throw new ServiceException("Error while adding user", e);
             }
@@ -210,16 +232,14 @@ public class UserServiceImpl implements UserService {
     }
 
     private boolean isEmailFree(String email) throws ServiceException {
-        UserValidator userValidator = new UserValidator();
         boolean result = false;
-        if (userValidator.isEmailValid(email)) {
-            try {
+        try {
+            if (UserValidator.isEmailValid(email)) {
                 Optional<User> optionalUser = userDao.findByEmail(email);
-                result = optionalUser.isPresent() ? false : true;
-            } catch (DaoException e) {
-                throw new ServiceException("Error while checking free email " +
-                        "for presence in database", e);
+                result = optionalUser.isPresent();
             }
+        } catch (DaoException e) {
+            throw new ServiceException("Error while checking free email for presence in database", e);
         }
         return result;
     }
